Could not make standard hadleclicks work. I finished seprating concerns and making components for each insert.
